translate mitochondrial variants using mitochondrial genetic code
